bluetooth: FakeBluetoothChooser impl.
Implements the interfaces that FakeBluetoothChooser overrides. The
mojom::FakeBluetoothChooser overrides are implemented by running
|event_handler_| with the appropriate arguments. The BluetoothChooser
overrides do nothing at the moment, since the base class is only meant
to update the chooser user interface.

This change also begins to partially convert a Bluetooth chooser test to test
the interface and deletes the temporary tests that were created to test this
API.

BUG=719827

Change-Id: I56a04e6bdb2de38d9163d7bc350db578deeb9ba3